Parse FETCH response to BODY[HEADER...] request
When FETCHing (BODY[HEADER.FIELDS (...)]), the key of the response
contains embedded spaces.  Parse this properly by grouping the "BODY["
and the matching ']', returning the entire block as a string.

There are still valid imap responses that won't correctly be parsed by
this, but this allows for this common use case to work.